Hopefully this'll fix the registration issues

+++ bin/user/usermessage.sh
@@ -247,9 +247,11 @@ if [[ "${ignoreUser}" -eq "0" ]]; then
 			done
 			;;
 		MODE)
-			# Someone changed modes? Time for a new names!
-			rm "var/.track/${senderTarget,,}"
-			echo "NAMES ${senderTarget}" >> "${output}"
+			if [[ "${senderTarget:0:1}" =~ [#|&] ]]; then
+			# Someone changed modes in a channel? Time for a new names!
+				rm "var/.track/${senderTarget,,}"
+				echo "NAMES ${senderTarget}" >> "${output}"
+			fi
 			;;
 		PART) 
 			# MySQL Seen Stuff
